update eudat-v1.re so irods v3.3 will work with it because of 2 problems: 1 ERROR: executeRuleAction Failed for readFile status = -1230000 RE_TYPE_ERROR 2 segfault of the irodsAgent
The first problem was caused by wrong types. It tried to loop over a string.
This has to be done over a array

The second problem was caused by the return values of an "elem" function to be overwritten.
The "elem" function was not properly used. I was changed to an "foreach" loop.